refactor: rename robotoff.slack into robotoff.notifier
